requirements.txt
psycopg2==2.9.7
requests==2.23.0
termcolor==1.1.0
python-dotenv==0.15.0
deepdiff
pygnparser
-e git+https://github.com/iobis/pyworms.git#egg=pyworms